Make deinterlace automatic for deinterlaced streams.
The video decoder does know if a video stream is interlaced, but only after decoding the first frame.
This requires quite a big reorganisation to do most of the openmax setup after the first frame is decoded.
The PortSettingsChanged message is the correct time to do this.

Change the command line arguments to mean default is automatic
--deinterlace/-d is force deinterlace
--nodeinterlace  is force no deinterlace (avoids allocating extra buffers on gpu)